Replace wood stove related info by pellet stove one
Updated references as text as the stove is no longer using wood but pellet.

Remove images that were outdated showing just a wood stove. In the future they
will be replaced.

NOTE: These changes are for keeping the record in source code but it is not
compiled to the output code. For the time required and effort it is just
easier to change source code to know what to touch in minified one.